refactor: 주석 처리 및 더미데이터 파일로 이동
