+++ constants/dummyData.ts
@@ -1,4 +1,4 @@
-import type { CreateGroupRequest, Grouping } from '@/apis/groups/type';
+import type { Article, CreateGroupRequest, Grouping } from '@/apis/groups';
 
 export const DUMMY_GROUPING_DATA: Grouping[] = [
   {
@@ -46,7 +46,7 @@ export const DUMMY_GROUPING_DATA: Grouping[] = [
 ];
 
 export const CREATE_GROUP_DUMMY_DATA: CreateGroupRequest = {
-  fileUrl: '그룹 이미지 Url',
+  imageUrl: '그룹 이미지 Url',
   title: '김지환이의 그룹',
   content: '안녕하세요',
   meetDate: '2022-09-07',
@@ -72,3 +72,38 @@ export const DUMMY_SIGN_UP_DATA = {
   gender: 'MAIL',
   personalities: ['OUTGOING'],
 };
+
+export const DUMMY_DATA: Article[] = [
+  {
+    articleId: 1,
+    images: [],
+    notice: true,
+    userImageUrl: '/assets/avatar.svg',
+    name: 'Kim',
+    date: '2021.09.01',
+    content:
+      '안녕하세요! 모임에 가입해주셔서 감사드립니다.\n오늘 19:00에 있을 모임 참여 전 모임에 필요한 공지 및 전달사항 확인 부탁드리겠습니다.',
+    commentCount: 0,
+  },
+  {
+    articleId: 2,
+    images: [],
+    notice: false,
+    userImageUrl: '/assets/avatar.svg',
+    name: 'Kim',
+    date: '2021.09.01',
+    content: '안녕하세요! 모임에 가입해주셔서 감사드립니다.',
+    commentCount: 100,
+  },
+  {
+    articleId: 3,
+    images: [],
+    notice: false,
+    userImageUrl: '/assets/avatar.svg',
+    name: 'Kim',
+    date: '2021.09.01',
+    content:
+      '안녕하세요! 모임에 가입해주셔서 감사드립니다.\n오늘 19:00에 있을 모임 참여 전 모임에 필요한 공지 및 전달사항 확인 부탁드리겠습니다.',
+    commentCount: 0,
+  },
+];
